feat: use Architectury to support different modloaders
