+++ .gitignore
@@ -1,3 +1,5 @@
+.architectury-transformer
+
 # gradle
 
 .gradle/
