+++ common/src/main/java/dev/ascpixi/dpf/DisconnectPacketFix.java
@@ -1,16 +1,15 @@
 package dev.ascpixi.dpf;
 
-import net.fabricmc.api.ModInitializer;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class DisconnectPacketFix implements ModInitializer
+public final class DisconnectPacketFix
 {
+    public static final String MOD_ID = "disconnect-packet-fix";
+
     public static final Logger LOGGER = LoggerFactory.getLogger("disconnect-packet-fix");
 
-    @Override
-    public void onInitialize()
+    public static void init()
     {
         LOGGER.info("Disconnect Packet Fix has started.");
     }
