+++ common/src/main/resources/disconnect-packet-fix.mixins.json
@@ -1,0 +1,14 @@
+{
+  "required": true,
+  "package": "dev.ascpixi.dpf.mixin",
+  "compatibilityLevel": "JAVA_21",
+  "minVersion": "0.8",
+  "client": [
+  ],
+  "mixins": [
+    "ConnectionMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
