--- fabric/gradle.properties
@@ -1,17 +1,0 @@
-# Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
-org.gradle.parallel=true
-
-# Fabric Properties
-# check these on https://fabricmc.net/develop
-minecraft_version=1.21
-yarn_mappings=1.21+build.7
-loader_version=0.15.11
-
-# Mod Properties
-mod_version=1.0.0
-maven_group=dev.ascpixi.dpf.ascpixi
-archives_base_name=disconnect-packet-fix
-
-# Dependencies
-fabric_version=0.100.4+1.21
