--- fabric/src/main/resources/disconnect-packet-fix.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"package": "dev.ascpixi.dpf.mixin",
-	"compatibilityLevel": "JAVA_21",
-	"mixins": [
-		"ConnectionMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
