+++ gradle.properties
@@ -1,0 +1,17 @@
+# Done to increase the memory available to Gradle.
+org.gradle.jvmargs=-Xmx2G
+org.gradle.parallel=true
+
+# Mod properties
+mod_version = 2.0.0
+maven_group = dev.ascpixi.dpf
+archives_name = disconnect-packet-fix
+enabled_platforms = fabric,neoforge
+
+# Minecraft properties
+minecraft_version = 1.21
+
+# Dependencies
+fabric_loader_version = 0.16.9
+fabric_api_version = 0.102.0+1.21
+neoforge_version = 21.0.42-beta
