+++ neoforge/gradle.properties
@@ -1,0 +1,1 @@
+loom.platform = neoforge
