+++ settings.gradle
@@ -1,0 +1,14 @@
+pluginManagement {
+    repositories {
+        maven { url "https://maven.fabricmc.net/" }
+        maven { url "https://maven.architectury.dev/" }
+        maven { url "https://files.minecraftforge.net/maven/" }
+        gradlePluginPortal()
+    }
+}
+
+rootProject.name = 'disconnect-packet-fix'
+
+include 'common'
+include 'fabric'
+include 'neoforge'
